+++ src/runnable/int/instr.rs
@@ -6,9 +6,9 @@ pub enum Instr {
     /// Remove from the current memory cell.
     Decr(u8),
     /// Shift the data pointer to the right.
-    Next(usize),
+    Next(u16),
     /// Shift the data pointer to the left.
-    Prev(usize),
+    Prev(u16),
     /// Display the current memory cell as an ASCII character.
     Print,
     /// Read one character from stdin.
@@ -16,13 +16,13 @@ pub enum Instr {
     /// Set a value for the current cell.
     Set(u8),
     /// Add the current cell to the cell n spaces away and set the current cell to 0.
-    AddTo(isize),
+    AddTo(i16),
     /// Subtract the current cell from the cell n spaces away and set the current cell to 0.
-    SubFrom(isize),
+    SubFrom(i16),
     /// Multiply current cell by a factor and add to cell at offset, then set current to 0.
-    MultiplyAddTo(isize, u8),
+    MultiplyAddTo(i16, u8),
     /// Copy current cell to multiple offsets, then set current to 0.
-    CopyTo(Vec<isize>),
+    CopyTo(Vec<i16>),
     /// If the current memory cell is 0, jump forward by the contained offset.
     BeginLoop(usize),
     /// If the current memory cell is not 0, jump backward by the contained offset.
